moving DatasetParticipantCountResponse and DatasetParticipantCountRequest back to DatasetBuilderUtils
